Optimize day 19
This makes the canonicalization more discriminating by tracking a sign
along with the absolute coordinates. Moreover, by considering the beacon
offsets for a scan as an entire sorted list, we can speed up the matching
process dramatically (however this makes the algorithm heuristic instead
of strictly following the allowed assumption).